upd 0.0.1
git-svn-id: https://fv3.googlecode.com/svn/trunk@46 1faddcec-d5ef-11de-96bf-fbce4fd84a6e

+++ src/fv3/Bounds.java
@@ -30,7 +30,7 @@ public class Bounds
     public final Vector min, max;
 
 
-    public Bounds(float[][] b){
+    public Bounds(double[][] b){
         this(new Vector(b[0]),new Vector(b[1]));
     }
     public Bounds(Vector min, Vector max){
